+++ Patches/TESTING.cs
@@ -1,29 +1,16 @@
-﻿using BepInEx.Logging;
+﻿/*using BepInEx.Logging;
+using GameNetcodeStuff;
 using HarmonyLib;
-using LethalLib.Extras;
-using LethalLib.Modules;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using UnityEngine;
-using static SCP956.Plugin;
-using LethalLib;
-using static LethalLib.Modules.Enemies;
+using Unity.Collections.LowLevel.Unsafe;
 using Unity.Netcode;
-using GameNetcodeStuff;
-using static UnityEngine.ParticleSystem.PlaybackState;
-using Unity.Mathematics;
 using UnityEngine;
 using UnityEngine.UIElements;
-using Unity.Burst.Intrinsics;
-using System.Collections;
-using static UnityEngine.VFX.VisualEffectControlTrackController;
-using UnityEngine.InputSystem.Utilities;
 
 namespace SCP956.Patches
 {
-    // spawnpositiontypes: GeneralItemClass, TabletopItems, SmallItems
     [HarmonyPatch]
     internal class TESTING : MonoBehaviour
     {
@@ -31,72 +18,215 @@ namespace SCP956.Patches
 
         private static PlayerControllerB localPlayer { get { return StartOfRound.Instance.localPlayerController; } }
 
-        [HarmonyPostfix, HarmonyPatch(typeof(HUDManager), "PingScan_performed")]
+        static bool invis = false;
+        static bool noSpawning = false;
+        static bool godMode = false;
+
+        static List<string> commands = new List<string> { "/spawn", "/spawnitem", "/invis", "/teleport", "/nospawning", "/money", "/godmode", "/damage", "/refresh", "/lights", "/enemies", "/players", "/items", "/turret" };
+
+        [HarmonyPostfix, HarmonyPatch(typeof(HUDManager), nameof(HUDManager.PingScan_performed))]
         public static void PingScan_performedPostFix()
         {
-            /*foreach (PlayerControllerB player in FrozenPlayers)
+
+        }
+
+        [HarmonyPrefix, HarmonyPatch(typeof(HUDManager), nameof(HUDManager.SubmitChat_performed))]
+        public static void SubmitChat_performedPrefix(HUDManager __instance)
+        {
+            string msg = __instance.chatTextField.text;
+            string[] args = msg.Split(" ");
+            logger.LogDebug(msg);
+
+
+            // Comment these out
+            if (args[0] == "/help")
             {
-                logger.LogDebug(player.playerUsername);
-            }*/
+                foreach (var command in commands)
+                {
+                    logger.LogDebug(command);
+                }
+            }
+            if (args[0] == "/spawn")
+            {
+                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
+                int index = RoundManager.Instance.currentLevel.Enemies.FindIndex(x => x.enemyType.enemyName == args[1]);
+                RoundManager.Instance.SpawnEnemyOnServer(pos, UnityEngine.Random.Range(0f, 360f), index);
+                HUDManager.Instance.DisplayTip("Testing", $"Spawned enemy: {args[1]}");
+                RoundManager.Instance.RefreshEnemiesList();
+            }
+            if (args[0] == "/spawnitem")
+            {
+                args[1].Replace("_", " ");
+                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
+                Item item = StartOfRound.Instance.allItemsList.itemsList.Where(x => x.itemName == args[1]).FirstOrDefault();
+                if (item == null) { return; }
 
-            //GameObject burningParticlesContainer = Resources.FindObjectsOfTypeAll<GameObject>().Where(x => x.name == "burningParticlesContainer").FirstOrDefault();
-            //logger.LogDebug("burningParticlesContainer: " + burningParticlesContainer == null);
+                GameObject obj = UnityEngine.Object.Instantiate(item.spawnPrefab, pos, Quaternion.identity, StartOfRound.Instance.propsContainer);
+                obj.GetComponent<NetworkObject>().Spawn();
+                HUDManager.Instance.DisplayTip("Testing", $"Spawned item: {item.itemName}");
+            }
+            if (args[0] == "/invis")
+            {
+                if (args[1] == "true")
+                {
+                    invis = true;
+                }
+                else if (args[1] == "false")
+                {
+                    invis = false;
+                }
+                HUDManager.Instance.DisplayTip("Testing", $"Invis set to {invis}");
+            }
+            if (args[0] == "/teleport")
+            {
+                if (args[1] == "main")
+                {
+                    Vector3 pos = RoundManager.FindMainEntrancePosition(true, true);
+                    localPlayer.TeleportPlayer(pos);
+                }
+            }
+            if (args[0] == "/nospawning")
+            {
+                if (args[1] == "true")
+                {
+                    noSpawning = true;
+                }
+                else if (args[1] == "false")
+                {
+                    noSpawning = false;
+                }
+                HUDManager.Instance.DisplayTip("Testing", $"Enemy spawning set to {!noSpawning}");
+            }
+            if (args[0] == "/money")
+            {
+                int amount = Convert.ToInt32(args[1]);
+                FindObjectOfType<Terminal>().groupCredits += amount;
+                HUDManager.Instance.DisplayTip("Testing", $"Credits added: {amount}");
+            }
+            if (args[0] == "/godmode")
+            {
+                if (args[1] == "true")
+                {
+                    godMode = true;
+                }
+                else if (args[1] == "false")
+                {
+                    godMode = false;
+                }
+                HUDManager.Instance.DisplayTip("Testing", $"Godmode set to {godMode}");
+            }
+            if (args[0] == "/damage")
+            {
+                int amount = Convert.ToInt32(args[1]);
+                localPlayer.DamagePlayer(amount);
+                HUDManager.Instance.DisplayTip("Testing", $"Damaged self: {amount}");
+            }
+            if (args[0] == "/refresh")
+            {
+                RoundManager.Instance.RefreshEnemiesList();
+                RoundManager.Instance.RefreshEnemyVents();
+                RoundManager.Instance.RefreshLightsList();
+            }
+            if (args[0] == "/lights")
+            {
+                bool value = Convert.ToBoolean(args[1]);
+                RoundManager.Instance.TurnOnAllLights(value);
+                HUDManager.Instance.DisplayTip("Testing", $"Lights set to {value}");
+            }
+            if (msg == "/enemies")
+            {
+                logger.LogDebug("ENEMIES:");
+                foreach (var enemy in GetEnemies())
+                {
+                    logger.LogDebug(enemy.enemyType.enemyName);
+                }
+            }
+            if (msg == "/players")
+            {
+                logger.LogDebug("PLAYERS:");
+                foreach (var player in StartOfRound.Instance.allPlayerScripts)
+                {
+                    logger.LogDebug(player.playerUsername);
+                }
+            }
+            if (msg == "/items")
+            {
+                logger.LogDebug("ITEMS:");
+                foreach (var item in StartOfRound.Instance.allItemsList.itemsList)
+                {
+                    logger.LogDebug(item.itemName);
+                }
+            }
+            if (msg == "/turret")
+            {
+                Vector3 pos = localPlayer.transform.forward * 2f + localPlayer.transform.position;
+                GameObject turretObj = UnityEngine.Object.Instantiate(StartOfRound.Instance.levels.Where(x => x.PlanetName == "41 Experimentation").First().spawnableMapObjects.Where(x => x.prefabToSpawn.name == "TurretContainer").First().prefabToSpawn, pos, Quaternion.identity, RoundManager.Instance.mapPropsContainer.transform);
+                turretObj.GetComponent<NetworkObject>().Spawn(true);
+                HUDManager.Instance.DisplayTip("Testing", "Spawned Turret");
+            }
+        }
 
-            /*CandyBehavior candy = localPlayer.currentlyHeldObjectServer.gameObject.GetComponent<CandyBehavior>();
+        [HarmonyPrefix, HarmonyPatch(typeof(EnemyAI), nameof(EnemyAI.PlayerIsTargetable))]
+        public static bool PlayerIsTargetablePrefix(ref bool __result)
+        {
+            if (NetworkManager.Singleton.IsServer || NetworkManager.Singleton.IsHost)
+            {
+                if (invis)
+                {
+                    __result = false;
+                    return false;
+                }
+            }
+            return true;
+        }
 
-            if (candy == null) { return; }
-            logger.LogDebug("Pinata candy: " + candy.pinataCandy);*/
-            
-            //StatusEffectController.Instance.TransformPlayer(localPlayer);
+        [HarmonyPrefix, HarmonyPatch(typeof(RoundManager), nameof(RoundManager.SpawnEnemiesOutside))]
+        public static bool SpawnEnemiesOutsidePrefix()
+        {
 
-            //Vector3 pos = RoundManager.Instance.GetRandomNavMeshPositionInBoxPredictable(scp.transform.position, config956TeleportRange.Value, RoundManager.Instance.navHit, RoundManager.Instance.AnomalyRandom);
-            //scp.Teleport(pos);
+            if (noSpawning)
+            {
+                return false;
+            }
+            return true;
+        }
 
-            // deathAnimations:
-            // 0 = normal
-            // 1 = decapitation
-            // 2 = coilhead decapitation
-            // 3 = seizure
-            // 4 = disapearance // was masked
-            // 5 = mask
-            // 6 = burn
-            // spawnpositiontypes: GeneralItemClass, TabletopItems, SmallItems
-            // PlayerAge = 10;
-            //logger.LogDebug("ping scan performed");
-            //logger.LogDebug("Player Age: " + PlayerAge);
-            //logger.LogDebug("Hands?: " + SCP330Behavior.noHands);
-            //logger.LogDebug("Candy taken: " + SCP330Behavior.candyTaken);
+        [HarmonyPrefix, HarmonyPatch(typeof(RoundManager), nameof(RoundManager.SpawnInsideEnemiesFromVentsIfReady))]
+        public static bool SpawnInsideEnemiesFromVentsIfReadyPrefix()
+        {
 
-            //localPlayer.bodyParts[1].GetComponent<Renderer>().enabled = false; // TODO: Do something with this to hide the players hands?
-            //WhiteSpike — Today at 8:09 PM
-            //You would probably have to mess with what's shown in the camera (the local body) and then the model itself when viewed by others (which means executed by other clients)
-            /*[Debug: Pinata] spine.004
-            [Debug: Pinata] arm.R_lower
-            [Debug: Pinata] arm.L_lower
-            [Debug: Pinata] shin.R
-            [Debug: Pinata] shin.L
-            [Debug: Pinata] spine.002
-            [Debug: Pinata] Player
-            [Debug: Pinata] thigh.R
-            [Debug: Pinata] thigh.L
-            [Debug: Pinata] arm.L_upper
-            [Debug: Pinata] arm.R_upper*/
+            if (noSpawning)
+            {
+                return false;
+            }
+            return true;
+        }
 
+        [HarmonyPrefix, HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.DamagePlayer))]
+        public static bool DamagePlayerPrefix(int damageNumber)
+        {
+            if (godMode)
+            {
+                HUDManager.Instance.DisplayTip("Testing", $"Godmode prevented {damageNumber} damage to player");
+                return false;
+            }
+            return true;
+        }
 
-            //string effects = "status negation:10,true;DamageReduction:15, 35, false, true;HealthRegen:10,5;restorestamina:50;IncreasedMovementSpeed:15,10;";
-            //StatusEffectController.Instance.ApplyCandyEffects(effects);
-            //StatusEffectController.Instance.RestoreStamina(10);
-            //StatusEffectController.Instance.InfiniteSprint(10, true);
-            //StatusEffectController.Instance.IncreasedMovementSpeed(10, 10, true, true);
+        public static List<SpawnableEnemyWithRarity> GetEnemies()
+        {
+            logger.LogDebug("Getting enemies");
+            List<SpawnableEnemyWithRarity> enemies = new List<SpawnableEnemyWithRarity>();
+            enemies = GameObject.Find("Terminal")
+                .GetComponentInChildren<Terminal>()
+                .moonsCatalogueList
+                .SelectMany(x => x.Enemies.Concat(x.DaytimeEnemies).Concat(x.OutsideEnemies))
+                .Where(x => x != null && x.enemyType != null && x.enemyType.name != null)
+                .GroupBy(x => x.enemyType.name, (k, v) => v.First())
+                .ToList();
 
-            /*Item scp330 = LethalLib.Modules.Items.LethalLibItemList.Where(x => x.name == "CandyBowlItem").First();
-            Item scp330p = LethalLib.Modules.Items.LethalLibItemList.Where(x => x.name == "CandyBowlPItem").First();
-            logger.LogDebug("Got items");
-
-            logger.LogDebug(scp330.name + ": " + scp330.spawnPositionTypes.Count);
-            logger.LogDebug(scp330p.name + ": " + scp330p.spawnPositionTypes.Count);
-            //StartOfRound.Instance.ManuallyEjectPlayersServerRpc();
-            //RoundManager.Instance.SpawnScrapInLevel();*/
+            logger.LogDebug($"Enemy types: {enemies.Count}");
+            return enemies;
         }
     }
-}
+}*/